requirements.txt
fastapi==0.63.0
uvicorn[standard]==0.13.2
notion==0.0.27
black==20.8b1
PyYAML==5.3.1
pydantic==1.7.3